requirements.txt
streamlit
scikit-learn
joblib
sentence-transformers
python-docx
PyMuPDF
docx2txt
openai